Add files via upload
Hello!  I've added the server.py file which contains all of the API information, queries, to get and display the book information to the "books.html" template.  The books.html template is a boiler plate with the code needed to display book information. 

Also, as you'll notice in the "server.py" file, the API runs every time the "/books" route is loaded.  Of course, that can be changed or edited however the group sees fit.  We'll also have to pip install "requests" in the virtual environment in order to get the api to work.   Also, I"m very new to Git so hopefully I'm going this correctly!  I'll post the link to the "books" database table in the mm channel!  I'